+++ service/internal/indexcache/spec.go
@@ -1,0 +1,7 @@
+package indexcache
+
+import "context"
+
+type Interface interface {
+	GetIndex(ctx context.Context, url string) (*Index, error)
+}