+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Get tarball URL for chart CRs from index.yaml for better community app catalog support.
+
 ## [5.5.2] - 2022-01-28
 
 ### Fixed